Move TermThemeRenderer and the confirmation prompt to crossterm.
General things of note
- As crossterm operates on any io::Write trait object instead of
a dedicated Term struct, we use the static terminal::size()
method to obtain the terminal size, which returns an option.
- We conciously only support clearing an u16 number of lines, since that's
crossterm only allows moving the cursor by that many lines.
- We conciously manually clear the last N lines instead of the entire terminal
since the confirmation prompt expects it so.

This change was manually tested, by running the `confirm` example with
various inputs. The confirmation prompt behaves the same as before this change.